optimize translate
